refactor: improve ADC latency with prescaler setting

+++ src/display.h
@@ -14,8 +14,8 @@ static constexpr uint8_t GLYPH_RBRAK = 0xAE;
 static constexpr uint8_t DISPLAY_I2C_ADDR = 0x3C;
 static constexpr uint8_t BLINK_DELAY = 200;
 
-void drawMsValue(float mean_ms);
-void drawStdDevValue(float sd_ms);
+void drawMsValue(double mean_ms);
+void drawStdDevValue(double sd_ms);
 void printMeasurement(uint16_t baseline, uint8_t cycle_index, double cycle_latency = 0.0, uint16_t measured = 0);
 void printAverage(double mean_ms, double sd_ms);
 void showHeader();
